Redesign the building model
It is turned that we have two critical problems in the current design:

a. Structure building commands are not issued immediately, so we can't
rely on the current minerals/vespene values. As a result construction
fails.

b. Some workers receive the same tasks twice, because it takes several steps
for a unit to receive an order. As a result same worker is assigned to
different construction orders.

This commit is the major redesign.

To address #a we now reserve resources until a structure is actually
placed on the ground.

To address #b we track internally busy and free workers and currently
assigned jobs.

Also some renaming and refactoring were done.